+++ src/components/common/Img.tsx
@@ -1,23 +1,32 @@
-import styled from "styled-components";
+import styled, { StyledProps, css } from "styled-components";
 
-interface ImgProps {
+export interface ImageProps {
   src: string;
   alt: string;
   role?: string;
-  width?: number;
-  height?: number;
-  id?: string;
-  className?: string;
+  width?: number | string;
+  height?: number | string;
+  mode?: ImageModes;
 }
 
-const Image = styled.img`
-  display: block;
-  max-width: 100%;
-  height: auto;
-`;
+/** stretch sets 100% width on the img, inherit or undefined uses only max-width: 100% */
+export type ImageModes = "stretch" | "inherit";
 
-const Img: React.FC<ImgProps> = ({ src, alt, ...props }) => (
-  <Image src={src} alt={alt} {...props} loading="lazy" />
+const Image = styled.img(
+  ({ $mode }: StyledProps<{ $mode?: ImageModes }>) => css`
+    display: block;
+    height: auto;
+    max-width: 100%;
+
+    ${$mode === "stretch" &&
+    css`
+      width: 100%;
+    `}
+  `
+);
+
+const Img: React.FC<ImageProps> = ({ src, alt, mode, ...props }) => (
+  <Image src={src} alt={alt} $mode={mode} loading="lazy" {...props} />
 );
 
 export default Img;
